Keep old schema for avro / big query, adjust types in clickhouse store

+++ logscore/logscore.go
@@ -6,16 +6,19 @@ import (
 	"fmt"
 )
 
-// LogScore is the data structure for the 'log_score' table (monitoring measurements)
+// LogScore is the data structure for the 'log_scores' table
+// with monitoring measurements).
+// Many of the fields have more precision here than in the database,
+// this is kept to fit the avro / bigquery schema (int64 vs uint32).
 type LogScore struct {
 	ID        int64            `json:"id" msgpack:"id"`
-	ServerID  uint32           `json:"sid" msgpack:"sid"`
-	MonitorID uint32           `json:"mid" msgpack:"mid"`
+	ServerID  int64            `json:"sid" msgpack:"sid"`
+	MonitorID int64            `json:"mid" msgpack:"mid"`
 	Ts        int64            `json:"ts" msgpack:"ts"`
-	Score     float32          `json:"sc" msgpack:"sc"`
-	Step      float32          `json:"st" msgpack:"st"`
+	Score     float64          `json:"sc" msgpack:"sc"`
+	Step      float64          `json:"st" msgpack:"st"`
 	Offset    *float64         `json:"of" msgpack:"of"`
-	RTT       *uint32          `json:"rtt" msgpack:"rtt"`
+	RTT       *int64           `json:"rtt" msgpack:"rtt"`
 	Meta      LogScoreMetadata `json:"attributes,omitempty"`
 }
